YC: add product visibility
So we can change where is the product going to be visible (POS, eshop or both). The product still has to be activated first.
