+++ src/ya-comiste-backoffice/pages/_app.css
@@ -1,0 +1,17 @@
+* {
+  box-sizing: border-box;
+}
+
+html,
+body {
+  background: #fff;
+  color: #1c1e21;
+  direction: ltr;
+  font-family: Helvetica, Arial, sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  font-size: 1rem;
+}
