+++ src/ya-comiste-rust/server/src/commerce/dal/mod.rs
@@ -1,1 +1,4 @@
 pub(in crate::commerce) mod products;
+
+#[cfg(test)]
+mod tests;
